+++ testData/org/elixir_lang/reference/module/as/issue_2446/asd_xS@ps.ex
@@ -1,0 +1,4 @@
+defmodule :asd_xS@ps do
+  def foo do
+  end
+end